requirements.txt
cvxpylayer
matplotlib==3.1.3
numpy==1.18.1
scipy==1.4.1
-e git+https://github.com/pytorch/pytorch@5e27de021e355cdb77693f3ff63088bb75678d61#egg=torch
